Add files via upload
I download the S&P500 price data from yfinance and use my custom-built Bollinger Bands library to generate the upper and lower bands of the Bollinger Bands. Then, I take the upper and lower bands of the Bollinger Bands as features (X) and the stock price as the target variable (Y). By aligning the Bollinger Bands' upper and lower bands today with the stock price the following day in chronological order, I split the data into a training set and a test set using a 6.5:3.5 ratio. Then, I build a regression model using sklearn to predict stock prices, achieving an accuracy of approximately 97%. However, it's essential to note that this is just a reference and not suitable for determining entry and exit points in the market.